+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/tabs/behavior-tab/behaviors/report/simulated-region-overview-behavior-report.component.ts
@@ -98,6 +98,28 @@ export class SimulatedRegionOverviewBehaviorReportComponent implements OnInit {
         });
     }
 
+    updateReportTransferOfCategoryInSingleRegionCompleted(
+        reportsEnabled: boolean
+    ) {
+        this.exerciseService.proposeAction({
+            type: '[ReportBehavior] Update report transfer of category in single region completed',
+            simulatedRegionId: this.simulatedRegionId,
+            behaviorId: this.reportBehaviorId,
+            reportChanges: reportsEnabled,
+        });
+    }
+
+    updateReportTransferOfCategoryInMultipleRegionsCompleted(
+        reportsEnabled: boolean
+    ) {
+        this.exerciseService.proposeAction({
+            type: '[ReportBehavior] Update report transfer of category in multiple regions completed',
+            simulatedRegionId: this.simulatedRegionId,
+            behaviorId: this.reportBehaviorId,
+            reportChanges: reportsEnabled,
+        });
+    }
+
     updateInterval(informationType: ReportableInformation, interval: string) {
         this.exerciseService.proposeAction({
             type: '[ReportBehavior] Update Recurring Report',