+++ shared/src/models/radiogram/utils/scope.ts
@@ -1,8 +1,0 @@
-import type { AllowedValues } from '../../../utils/validators';
-
-export type Scope = 'singleRegion' | 'transportManagement';
-
-export const scopeAllowedValues: AllowedValues<Scope> = {
-    singleRegion: true,
-    transportManagement: true,
-};