+++ shared/src/state-migrations/migration-functions.ts
@@ -22,7 +22,8 @@ import { updateEocLog3 } from './3-update-eoc-log';
 import { reportTreatmentStatusChanges30 } from './30-report-treatment-status-changes';
 import { improveLoadVehicleActivity31 } from './31-improve-load-vehicle-activity';
 import { removeIdFromEvents32 } from './32-remove-id-from-events';
-import { addCatchAllHospital33 } from './33-add-catch-all-hospital';
+import { reportTransferCategoryCompleted33 } from './33-report-transfer-category-completed';
+import { addCatchAllHospital34 } from './34-add-catch-all-hospital';
 import { removeSetParticipantIdAction4 } from './4-remove-set-participant-id-action';
 import { removeStatistics5 } from './5-remove-statistics';
 import { removeStateHistory6 } from './6-remove-state-history';
@@ -88,5 +89,6 @@ export const migrations: {
     30: reportTreatmentStatusChanges30,
     31: improveLoadVehicleActivity31,
     32: removeIdFromEvents32,
-    33: addCatchAllHospital33,
+    33: reportTransferCategoryCompleted33,
+    34: addCatchAllHospital34,
 };