+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project does **not** adhere to [Semantic Versioning](https://semver.org
     -   When a vehicle is selected, the corresponding material and personnel are highlighted as well.
 -   A generic catch-all hospital is present in every exercise and cannot be deleted.
 -   The reports behavior can generate reports on the counts of transferred patients per triage category.
+-   The reports behavior can generate event-based reports when the last patient of a triage category has been transferred to a hospital.
 
 ### Changed